requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
django-health-check==3.4.1
django-oauth-toolkit==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.7.7
dotenv==0.0.5
elasticache-auto-discovery==1.0.0
gensim==3.2.0
Logentries==0.17
numpy==1.13.3
pandas==0.22.0
pylint==1.8.1
pymemcache==1.4.4
pystan==2.17.0.0
python-dotenv==0.7.1
PyWavelets==0.5.2
scikit-learn==0.19.1
scipy==1.0.0
sklearn==0.0
statistics==1.0.3.5
